Put a lot of time thinking about issue #8
I've made some useful progress and have a solution.

Unfortunately, the solution infects other parts
of the system. The underlying cause is likely
around the use of ‘org-indent-region’, but there may
be more.

For now, I've left my explorations under the heading
“Issue ♯8”; which contains code that may eventually
migrate to the location where ‘defblock’ is implemented.

The system has gotten large and automated tests are
needed.